server/requirements.txt
annotated-doc==0.0.4
annotated-types==0.7.0
anyio==4.11.0
beanie==2.0.1
click==8.3.1
dnspython==2.8.0
fastapi==0.121.3
graphql-core==3.2.7
h11==0.16.0
idna==3.11
lazy-model==0.4.0
lia-web==0.2.3
motor==3.7.1
packaging==25.0
pydantic==2.12.4
pydantic_core==2.41.5
pymongo==4.15.4
python-dateutil==2.9.0.post0
python-dotenv==1.0.1
python-multipart==0.0.20
six==1.17.0
sniffio==1.3.1
starlette==0.50.0
strawberry-graphql==0.286.1
typing-inspection==0.4.2
typing_extensions==4.15.0
uvicorn==0.38.0
